backend/requirements.txt
Flask
Flask-CORS
python-dotenv
google-generativeai
numpy
faiss-cpu 
pandas
networkx
spacy
